Add script to sort images automatically using trained CNN
The script loads a pre-trained CNN classifier model. It loads the
label binarizer to find the class names.

It then iterates through the input folder and preprocess the images.

The given CNN model is used to classify the images. Images above a
prediction confidence threshold are moved to their matching class-
specific subfolder. Images below the threshold are skipped.